# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Finante API
  description: Personal finance management API for tracking transactions, labels, and rules
  version: 1.0.0
  contact:
    name: Finante API Support
    email: cosmin@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.finante.example.com/api/v1
    description: Production server
...
paths:
  /transactions:
    get:
      tags:
        - transactions
      summary: Get transactions
      description: Retrieve transactions with optional date filtering
      parameters:
        - name: startDate
          in: query
          description: Start date as milliseconds since epoch
          required: false
          schema:
            type: integer
            format: int64
            example: 1704067200000
        - name: endDate
          in: query
          description: End date as milliseconds since epoch
          required: false
          schema:
            type: integer
            format: int64
            example: 1706745600000
      responses:
        '200':
          description: Successfully retrieved transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transactions'
        '400':
          description: Bad request - invalid date parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - transactions
      summary: Create transaction
      description: Create a new transaction
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionForm'
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Bad request - validation error or transaction already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /transactions/{id}:
    get:
      tags:
        - transactions
      summary: Get transaction by id
      description: Get an existing transaction
      operationId: getTransaction
      parameters:
        - name: id
          in: path
          required: true
          description: Transaction ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Transaction updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Bad request - invalid ID or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - transactions
      summary: Update transaction
      description: Update an existing transaction
      operationId: updateTransaction
      parameters:
        - name: id
          in: path
          required: true
          description: Transaction ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionForm'
      responses:
        '201':
          description: Transaction updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Bad request - invalid ID or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - transactions
      summary: Delete transaction
      description: Delete a transaction by ID
      operationId: deleteTransaction
      parameters:
        - name: id
          in: path
          required: true
          description: Transaction ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Transaction deleted successfully
        '400':
          description: Bad request - invalid ID or transaction cannot be deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /transactions/{id}/labels:
    get:
      tags:
        - transactions
      summary: Get transaction labels
      description: Get all labels associated with a transaction
      operationId: getTransactionLabels
      parameters:
        - name: id
          in: path
          required: true
          description: Transaction ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved transaction labels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Labels'
        '400':
          description: Bad request - invalid transaction ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - transactions
      summary: Add label to transaction
      description: Add a label to a transaction
      operationId: addTransactionLabel
      parameters:
        - name: id
          in: path
          required: true
          description: Transaction ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelForm'
      responses:
        '201':
          description: Label added to transaction successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Label'
        '400':
          description: Bad request - invalid transaction ID or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - transactions
      summary: Remove all labels from transaction
      description: Remove all labels from a transaction
      operationId: removeTransactionLabels
      parameters:
        - name: id
          in: path
          required: true
          description: Transaction ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: All labels removed from transaction successfully
        '400':
          description: Bad request - invalid transaction ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /transactions/{id}/labels/{labelId}:
    delete:
      tags:
        - transactions
      summary: Remove specific label from transaction
      description: Remove a specific label from a transaction
      operationId: removeTransactionLabel
      parameters:
        - name: id
          in: path
          required: true
          description: Transaction ID
          schema:
            type: integer
            format: int64
        - name: labelId
          in: path
          required: true
          description: Label ID
          schema:
            type: integer
      responses:
        '204':
          description: Label removed from transaction successfully
        '400':
          description: Bad request - invalid transaction ID or label ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    TransactionForm:
      type: object
      required:
        - kind
        - date
        - content
        - amount
        - account
      properties:
        kind:
          type: string
          enum: [credit, debit]
          description: Transaction type
        date:
          type: string
          format: date
          pattern: '^\d{2}/\d{2}/\d{4}$'
          example: "15/01/2024"
          description: Transaction date in DD/MM/YYYY format
        content:
          type: string
          description: Transaction description/content
          example: "Grocery shopping at SuperMart"
        amount:
          type: number
          format: float
          description: Transaction amount
          example: 45.67
        account:
          type: integer
          format: int64
          description: Account number
          example: 1234567890
    Transaction:
      type: object
      properties:
        href:
          type: string
          format: uri
          example: "/api/v1/transactions/123"
        kind:
          type: string
          example: "EXPENSE"
        date:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        account:
          type: integer
          format: int64
          example: 1234567890
        description:
          type: string
          example: "Grocery shopping at SuperMart"
        amount:
          type: number
          format: float
          example: 45.67
        labels:
          type: array
          items:
            $ref: '#/components/schemas/TransactionLabelAssociation'
    TransactionLabelAssociation:
      type: object
      properties:
        href:
          type: string
          format: uri
          example: "/api/v1/labels/456"
        key:
          type: string
          example: "category"
        value:
          type: string
          example: "groceries"
        rule_href:
          type: string
          format: uri
          example: "/api/v1/rules/grocery-rule"
          description: Reference to the rule that created this label association
    Transactions:
      type: object
      properties:
        total:
          type: integer
          example: 150
        start:
          type: string
          format: date
          pattern: '^\d{2}/\d{2}/\d{4}$'
          example: "01/01/2024"
        end:
          type: string
          format: date
          pattern: '^\d{2}/\d{2}/\d{4}$'
          example: "31/01/2024"
        items:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    LabelForm:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
          description: Label key
          example: "category"
        value:
          type: string
          description: Label value
          example: "groceries"
    Label:
      type: object
      properties:
        href:
          type: string
          format: uri
          example: "/api/v1/labels/123"
        key:
          type: string
          example: "category"
        value:
          type: string
          example: "groceries"
        rules:
          type: array
          items:
            $ref: '#/components/schemas/Rule'
          description: Rules that can create this label
    Labels:
      type: object
      properties:
        total:
          type: integer
          example: 25
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
...
    Rule:
      type: object
      properties:
        href:
          type: string
          format: uri
          example: "/api/v1/rules/grocery-rule"
        name:
          type: string
          example: "grocery-rule"
        pattern:
          type: string
          example: "(?i).*(grocery|supermarket).*"
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
          description: Labels that this rule applies
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Validation failed"